agent/requirements.txt
langchain
langchain-openai
openai
python-dotenv
requests
tweepy
coinbase # Add if used